Consolidate search hooks
